Fix commit. Stuff done:
Better validating of user input.
Fixed styling in calendar
Fix bug with "0 anmälda" in calendar and admin
Added so its possible to register to event via calendar
Added last name to user. (new database)
Fixed "re-sending" forms bug
Removed "Användare <name>" to only <name>
Bug with preview only shown for logged in user

+++ include/src/Objects/DataObject.php
@@ -143,9 +143,14 @@ abstract class DataObject
 
 	public function removeSingleEntryById($id)
 	{
-		$sql = "DELETE FROM " . $this->table . " WHERE id=? LIMIT 1";
-		$this->database->ExecuteQuery($sql, array($id));
-		return true;
+		if(is_numeric($id))
+		{
+			$sql = "DELETE FROM " . $this->table . " WHERE id=? LIMIT 1";
+			$this->database->ExecuteQuery($sql, array($id));
+			return true;
+		}
+		return false;
+
 	}
 
 	public function rowCount()
